Updating readme and license; updating cpprestsdk
Updated the readme and license files to current standards
Updated the cpprestsdk to use the Microsft version. Removed information
about the fix that Cyrile had to do, as it is now in the latest
cpprestsdk.
Removed the fixed Casablanca files (cpprestsdk was formerly Casablanca)
Updated certain include paths to new cpprestsdk paths